[PATCH] Increase resolution for parameter automation
The float type was causing « rounding jitter » and couldn’t achieve sample-accurate automation

We use Float64 instead to be in line with the type of timestamp->mSampleTime

+++ Sources/AudioKitEX/Automation/NodeParameter+Automation.swift
@@ -47,8 +47,8 @@ extension NodeParameter {
             
             guard let observer = ParameterAutomationGetRenderObserver(parameter.address,
                                                                       avAudioNode.auAudioUnit.scheduleParameterBlock,
-                                                                      Float(Settings.sampleRate),
-                                                                      Float(lastTime.sampleTime),
+                                                                      Settings.sampleRate,
+                                                                      Float64(lastTime.sampleTime),
                                                                       automationBaseAddress,
                                                                       events.count) else { return }
